Support for additional CNN architectures.

+++ survos2/improc/utils.py
@@ -105,7 +105,7 @@ def optimal_chunksize(source, max_size, item_size=4, delta=0.1, axis_weight=None
         axis_weight = shape / float(np.min(shape))
 
     axis_weight = 1 + np.asarray(axis_weight, float) / np.max(axis_weight)
-    total_chunks = int(ceil(np.prod(shape).astype(np.float64) * item_size / sizeMB))
+    total_chunks = int(ceil(np.prod(shape).astype(np.float32) * item_size / sizeMB))
     total_chunks_axis = [int(ceil(total_chunks / p)) for p in axis_weight]
     max_chunk_iter = [range(1, total + 1) for total in total_chunks_axis]
     best_chunk = shape
